REFACTOR: Move `bind_arguments` to new module.
Also, rename the `inspect_names` module to `inspect_name`.
