+++ flowgraph/trace/inspect_name.py
